# qcom faster copies