sync:   13.08 ms @ 5.13 GB/s
vs
sync:    8.98 ms @ 7.47 GB/s

Now the same speed as GPU... WC is slower than WT and WB with memset. I think we might want NT memset (which bypasses caches?).